mpi first structure

+++ include/IO_VTK.hpp
@@ -15,11 +15,11 @@ template
 <typename T>
 void read_vtk(const std::string problem_name, std::vector<Vertex<T>>& vertices,
 				std::vector<Face>& faces,
-				size_t& nvertices, size_t& nfaces);
+				int& nvertices, int& nfaces);
 template
 <typename T>
 void write_vtu(const std::string problem_name, const std::vector<Vertex<T>>& vertices, const std::vector<Face>& faces,
-							size_t &nvertices, size_t &nfaces);
+							int &nvertices, int &nfaces);
 
 #endif
